Refactor: Выделение PDF-логики в `pdfGenerator.ts` и исправление отобр
Apply AI assistant changes to 2 files.
Ref: Refactor: Выделение PDF-логики в `pdfGenerator.ts` и исправление отображения Hero Image...
